+++ tox.ini
@@ -17,6 +17,7 @@ deps = coverage
 commands =
     coverage report
     coverage xml
+    coverage html
     coverage erase
